fix: move BasicAuth credentials to public package for external client use
