+++ pkg/client/auth/basic/credentials.go
@@ -1,4 +1,4 @@
-package common
+package basic
 
 import (
 	"context"
